misc precision fixes: float -> double, in some cases we need the decimals

+++ src/main/java/ch/digitalfondue/mjml4j/CssBoxModel.java
@@ -1,4 +1,4 @@
 package ch.digitalfondue.mjml4j;
 
-record CssBoxModel(float totalWidth, float borderWidth, float paddingWidth, float boxWidth) {
+record CssBoxModel(double totalWidth, double borderWidth, double paddingWidth, double boxWidth) {
 }
